tests/python_client/requirements.txt
--extra-index-url https://test.pypi.org/simple/
pytest-cov==2.8.1
requests>=2.32.4
scikit-learn>=1.5.2
ujson==5.5.0
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-timeout==1.3.3
pytest-repeat==0.8.0
allure-pytest==2.7.0
pytest-print==0.2.1
pytest-level==0.1.1
pytest-xdist==2.5.0

pytest-assume==2.4.3
pytest-rerunfailures==14.0
pytest_tagging==1.6.0
pytest-html==3.1.1
delayed-assert==0.3.5
kubernetes==17.17.0
PyYAML==6.0
pytest-sugar==0.9.5
pytest-random-order

# pymilvus
pymilvus==2.7.0rc141
pymilvus[bulk_writer]==2.7.0rc141
# for protobuf
protobuf>=5.29.5

# for customize config test
python-benedict==0.24.3

# for bulk insert test
minio==7.2.0
npy-append-array==0.9.15
Faker==19.2.0

# for benchmark
h5py==3.8.0

# for log
loguru==0.7.0

# util
psutil==5.9.4
pandas==1.5.3
numpy==1.26.4
tenacity==8.1.0
rich==13.7.0
# for standby test
etcd-sdk-python==0.0.6
deepdiff==8.6.1

# for test result analyzer
prettytable==3.8.0
pyarrow==14.0.1
fastparquet==2023.7.0

# for bf16 datatype
ml-dtypes==0.2.0

# for text processing
Unidecode==1.3.8

# for full text search
tantivy==0.22.0
bm25s==0.2.0
jieba==0.42.1


# for perf test
locust==2.25.0

# for supporting higher python version
typing_extensions==4.12.2

# for env configuration
python-dotenv<=2.0.0

# for geometry
shapely==2.1.1

# for time zone
tzdata>=2024.1
